+++ api/app/actors/ProjectActor.scala
@@ -4,16 +4,11 @@ import io.flow.dependency.api.lib._
 import io.flow.dependency.v0.models._
 import io.flow.dependency.v0.models.json._
 import io.flow.postgresql.Pager
-import io.flow.play.util.Config
 import db._
-import akka.actor.{Actor, ActorSystem}
+import akka.actor.Actor
 import io.flow.akka.SafeReceive
-import io.flow.github.v0.models.HookForm
 import io.flow.log.RollbarLogger
 import play.api.libs.json.Json
-import play.api.libs.ws.WSClient
-
-import scala.concurrent.ExecutionContext
 
 object ProjectActor {
 
@@ -21,9 +16,7 @@ object ProjectActor {
 
   object Messages {
 
-    case object Deleted extends Message
-    case object CreateHooks extends Message
-    case object Sync extends Message
+    case class Delete(projectId: String) extends Message
     case object SyncCompleted extends Message
 
     case class ProjectLibraryCreated(id: String) extends Message
@@ -45,10 +38,7 @@ object ProjectActor {
 
 class ProjectActor @javax.inject.Inject() (
   logger: RollbarLogger,
-  config: Config,
-  actorSystem: ActorSystem,
   projectsDao: ProjectsDao,
-  tokensDao: TokensDao,
   syncsDao: SyncsDao,
   projectBinariesDao:ProjectBinariesDao,
   projectLibrariesDao: ProjectLibrariesDao,
@@ -57,18 +47,12 @@ class ProjectActor @javax.inject.Inject() (
   binariesDao: BinariesDao,
   usersDao: UsersDao,
   resolversDao: ResolversDao,
-  wsClient: WSClient,
   @com.google.inject.assistedinject.Assisted projectId: String
 ) extends Actor {
 
   private[this] lazy val SystemUser = usersDao.systemUser
 
-  private[this] implicit val ec : ExecutionContext = actorSystem.dispatchers.lookup("project-actor-context")
   private[this] implicit val configuredRollbar: RollbarLogger = logger.fingerprint(getClass.getName)
-
-  private[this] val HookBaseUrl = config.requiredString("dependency.api.host") + "/webhooks/github/"
-  private[this] val HookName = "web"
-  private[this] val HookEvents = Seq(io.flow.github.v0.models.HookEvent.Push)
 
   private[this] lazy val dataProject: Option[Project] = projectsDao.findById(Authorization.All, projectId)
 
@@ -96,123 +80,11 @@ class ProjectActor @javax.inject.Inject() (
         processPendingSync(project)
       }
 
-    case ProjectActor.Messages.CreateHooks =>
-      dataProject.foreach { project =>
-        GithubUtil.parseUri(project.uri) match {
-          case Left(error) => {
-            logger.fingerprint(getClass.getName).withKeyValue("project", Json.toJson(project)).withKeyValue("error", error).warn("error creating github hooks")
-          }
-          case Right(repo) => {
-            tokensDao.getCleartextGithubOauthTokenByUserId(project.user.id) match {
-              case None => {
-                logger.fingerprint(getClass.getName).withKeyValue("project", Json.toJson(project)).warn(s"No oauth token for user")
-              }
-
-              case Some(token) => {
-                val client = GithubHelper.apiClient(wsClient, token)
-
-                client.hooks.get(repo.owner, repo.project).map { hooks =>
-                  val targetUrl = HookBaseUrl + project.id
-                  hooks.find(_.config.url.contains(targetUrl)).getOrElse {
-                    client.hooks.post(
-                      owner = repo.owner,
-                      repo = repo.project,
-                      hookForm = HookForm(
-                        name = HookName,
-                        config = io.flow.github.v0.models.HookConfig(
-                          url = Some(targetUrl),
-                          contentType = Some("json")
-                        ),
-                        events = HookEvents,
-                        active = true
-                      )
-                    ).map { _ =>
-                      ()
-                    }.recover {
-                      case e: Throwable => {
-                        logger.fingerprint(getClass.getName).withKeyValue("project", Json.toJson(project)).error("Error creating hook", e)
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-
-    case ProjectActor.Messages.Sync =>
-      dataProject.foreach { project =>
-        syncsDao.recordStarted("project", project.id)
-
-        val summary = projectsDao.toSummary(project)
-
-        GithubDependencyProviderClient.instance(wsClient, config, tokensDao, summary, project.user, logger).dependencies(project).map { dependencies =>
-          dependencies.binaries.foreach { binaries =>
-            val projectBinaries = binaries.map { form =>
-              projectBinariesDao.upsert(project.user, form) match {
-                case Left(errors) => {
-                  logger.withKeyValue("project", Json.toJson(project)).withKeyValue("form", form.toString).withKeyValue("errors", errors).error(s"Errors storing binary")
-                  None
-                }
-                case Right(projectBinary) => {
-                  Some(projectBinary)
-                }
-              }
-            }
-            projectBinariesDao.setIds(project.user, project.id, projectBinaries.flatten)
-          }
-
-          dependencies.librariesAndPlugins.foreach { libraries =>
-            val projectLibraries = libraries.map { artifact =>
-              val bins = dependencies.crossBuildVersion()
-              val crossBuildVersion = {
-                if (artifact.isPlugin)
-                  bins.get(BinaryType.Sbt).orElse(bins.get(BinaryType.Scala))
-                else
-                  bins.get(BinaryType.Scala)
-              }
-              projectLibrariesDao.upsert(
-                project.user,
-                artifact.toProjectLibraryForm(
-                  crossBuildVersion = crossBuildVersion
-                )
-              ) match {
-                case Left(errors) => {
-                  logger
-                    .withKeyValue("project", Json.toJson(project))
-                    .withKeyValue("errors", errors)
-                    .withKeyValue("artifact", artifact.toString)
-                    .warn("Validation errors storing artifact")
-                  None
-                }
-                case Right(library) => {
-                  Some(library)
-                }
-              }
-            }
-            projectLibrariesDao.setIds(project.user, project.id, projectLibraries.flatten)
-          }
-
-          recommendationsDao.sync(SystemUser, project)
-
-          processPendingSync(project)
-        }.recover {
-          case e => {
-            logger
-              .withKeyValue("project", Json.toJson(project))
-              .error("Error fetching dependencies", e)
-          }
-        }
-      }
-
-    case ProjectActor.Messages.Deleted =>
-      dataProject.foreach { project =>
-        Pager.create { offset =>
-          recommendationsDao.findAll(Authorization.All, projectId = Some(project.id), offset = offset)
-        }.foreach { rec =>
-          recommendationsDao.delete(SystemUser, rec)
-        }
+    case ProjectActor.Messages.Delete(projectId: String) =>
+      Pager.create { offset =>
+        recommendationsDao.findAll(Authorization.All, projectId = Some(projectId), offset = offset)
+      }.foreach { rec =>
+        recommendationsDao.delete(SystemUser, rec)
       }
       context.stop(self)
